src/main/java/lambdasinaction/chap5/Finding.java
package lambdasinaction.chap5;

import lambdasinaction.chap4.Dish;

import java.util.Arrays;
import java.util.List;
import java.util.Optional;

import static lambdasinaction.chap4.Restaurant.menu;

public class Finding {

  public static boolean isVegetarianFriendlyMenu() {
    return menu.stream().anyMatch(Dish::isVegetarian);
  }

  public static boolean isHealthyMenu() {
    return menu.stream().allMatch(d -> d.getCalories() < 1000);
  }

  public static boolean isHealthyMenu2() {
    return menu.stream().noneMatch(d -> d.getCalories() >= 1000);
  }

  public static Optional<Dish> findVegetarianDish() {
    return menu.stream().filter(Dish::isVegetarian).findAny();
  }

  public static Optional<Integer> findFirstSquareDivisibleByThree() {
    List<Integer> someNumbers = Arrays.asList(1, 2, 3, 4, 5);
    return someNumbers.stream()
        .map(x -> x * x)
        .filter(x -> x % 3 == 0)
        .findFirst();
  }
}
